backend/requirements-minimal.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
setuptools>=78.1.1
python-docx>=1.1.0
aiofiles>=23.2.1
aiosqlite>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
black>=23.10.0
ruff>=0.1.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
